RFDB: feito; FIT: data types alterados de unsigned long para uint32_t

+++ sofs19/src/grp_src/grp_mksofs/grp_mksofs_FIT.cpp
@@ -19,15 +19,14 @@ namespace sofs19
         /* change the following line by your code */
         
         // Número de blocks on Inode Table
-        unsigned long nBLocks = itotal/IPB;
-        unsigned long nInode = 0;
+        uint32_t nBLocks = itotal/IPB;
+        uint32_t nInode = 0;
         // Iterar pela tabela
-        for(unsigned long block=1; block<=nBLocks; block++) {
+        for(uint32_t block=1; block<=nBLocks; block++) {
             // Inicializar Bloco
             SOInode inodeBlock[IPB];
-            for(unsigned long inode=0; inode<IPB; inode++) {
+            for(uint32_t inode=0; inode<IPB; inode++) {
                 // Inicializa cada inode com valores default
-                //inodeBlock[inode].mode = 0175000;
                 inodeBlock[inode].mode = INODE_FREE;
                 inodeBlock[inode].lnkcnt = 0;
                 inodeBlock[inode].owner = 0;
@@ -37,7 +36,7 @@ namespace sofs19
                 inodeBlock[inode].ctime = 0;
                 inodeBlock[inode].mtime = 0;
                 inodeBlock[inode].next = ++nInode;
-                unsigned long i;
+                uint32_t i;
                 for(i=0; i<N_DIRECT; i++)
                     inodeBlock[inode].d[i] = NullReference;
                 for(i=0; i<N_INDIRECT; i++)
